# Conflicts:
#	src/main/java/com/mycompany/project/attendance/repository/AttendanceRepository.java
#	src/main/java/com/mycompany/project/course/service/CourseService.java
#	src/main/java/com/mycompany/project/reservation/query/service/ReservationService.java

+++ src/main/java/com/mycompany/project/reservation/command/entity/Reservation.java
@@ -17,7 +17,7 @@ import java.time.LocalDateTime;
 public class Reservation {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int reservationId;
+    private Long reservationId;
     private LocalDateTime startTime;
     private LocalDateTime endTime;
     private String status;
